Class in python file was corrected
